msmarco-v2-vector: include PR feedback from @jimczi
- Rename to msmarco-v2 / msmarco-v2-vector
- Use dot_product similarity for mapping
- Pull queries from https://ir-datasets.com/msmarco-passage-v2.html#msmarco-passage-v2/train
