--- cohere_v3_vector/.gitignore
@@ -1,1 +1,0 @@
-queries.json
