--- cohere_v3_vector/_tools/requirements.txt
@@ -1,2 +1,0 @@
-datasets
-cohere
