--- cohere_v3_vector/challenges/default.json
@@ -1,102 +1,0 @@
-{
-  "name": "index-and-search",
-  "description": "",
-  "default": true,
-  "schedule": [
-    {
-      "operation": {
-        "operation-type": "delete-index"
-      }
-    },
-    {
-      "name": "create-index",
-      "operation": "create-index"
-    },
-    {
-      "name": "check-cluster-health",
-      "operation": "check-cluster-health"
-    },
-    {
-      "name": "initial-documents-indexing",
-      "operation": "initial-documents-indexing",
-      "warmup-time-period": {{ initial_indexing_bulk_warmup | default(40) | int }},
-      "clients": {{ initial_indexing_bulk_indexing_clients | default(5) | int }}
-    },
-    {
-      "name": "refresh-after-index",
-      "operation": {
-        "operation-type": "refresh",
-        "request-timeout": 1000,
-        "include-in-reporting": true
-      }
-    },
-    {
-      "name": "wait-until-merges-finish-after-index",
-      "operation": {
-        "operation-type": "index-stats",
-        "index": "_all",
-        "condition": {
-          "path": "_all.total.merges.current",
-          "expected-value": 0
-        },
-        "retry-until-success": true,
-        "include-in-reporting": false
-      }
-    },
-    {# serverless-post-ingest-sleep-marker-start #}{%- if post_ingest_sleep|default(false) -%}
-    {
-      "name": "post-ingest-sleep",
-      "operation": {
-        "operation-type": "sleep",
-        "duration": {{ post_ingest_sleep_duration|default(30) }}
-      }
-    },
-    {%- endif -%}{# serverless-post-ingest-sleep-marker-end #}
-    {
-      "name": "standalone-search-knn-10-100-single-client",
-      "operation": "knn-search-10-100",
-      "warmup-iterations": 100,
-      "iterations": {{ standalone_search_iterations | default(10000) | int }}
-    },
-    {
-      "name": "standalone-knn-search-100-1000-single-client",
-      "operation": "knn-search-100-1000",
-      "warmup-iterations": 100,
-      "iterations": {{ standalone_search_iterations | default(10000) | int }}
-    },
-    {
-      "name": "standalone-search-knn-10-100-multiple-clients",
-      "operation": "knn-search-10-100",
-      "warmup-iterations": 100,
-      "clients": {{ standalone_search_clients | default(8) | int }},
-      "iterations": {{ standalone_search_iterations | default(10000) | int }}
-    },
-    {
-      "name": "standalone-search-knn-100-1000-multiple-clients",
-      "operation": "knn-search-100-1000",
-      "warmup-iterations": 100,
-      "clients": {{ standalone_search_clients | default(8) | int }},
-      "iterations": {{ standalone_search_iterations | default(10000) | int }}
-    },
-    {
-      "parallel": {
-        "tasks": [
-          {
-            "name": "parallel-documents-indexing-bulk",
-            "operation": "parallel-documents-indexing",
-            "clients": {{ parallel_indexing_bulk_clients | default(1) | int }},
-            "time-period": {{ parallel_indexing_time_period | default(1800) | int }},
-            "target-throughput": {{ parallel_indexing_bulk_target_throughput | default(1) | int }}
-          },
-          {
-            "name": "parallel-documents-indexing-search-knn-10-100",
-            "operation": "knn-search-10-100",
-            "clients": {{ parallel_indexing_search_clients | default(3) | int }},
-            "time-period": {{ parallel_indexing_time_period | default(1800) | int }},
-            "target-throughput": {{ parallel_indexing_search_target_throughput | default(100) | int }}
-          }
-        ]
-      }
-    }
-  ]
-}
