--- cohere_v3_vector/files.txt
@@ -1,188 +1,0 @@
-cohere-documents-01-1k.json
-cohere-documents-01-1k.json.bz2
-cohere-documents-01.json
-cohere-documents-01.json.bz2
-cohere-documents-02-1k.json
-cohere-documents-02-1k.json.bz2
-cohere-documents-02.json
-cohere-documents-02.json.bz2
-cohere-documents-03-1k.json
-cohere-documents-03-1k.json.bz2
-cohere-documents-03.json
-cohere-documents-03.json.bz2
-cohere-documents-04-1k.json
-cohere-documents-04-1k.json.bz2
-cohere-documents-04.json
-cohere-documents-04.json.bz2
-cohere-documents-05-1k.json
-cohere-documents-05-1k.json.bz2
-cohere-documents-05.json
-cohere-documents-05.json.bz2
-cohere-documents-06-1k.json
-cohere-documents-06-1k.json.bz2
-cohere-documents-06.json
-cohere-documents-06.json.bz2
-cohere-documents-07-1k.json
-cohere-documents-07-1k.json.bz2
-cohere-documents-07.json
-cohere-documents-07.json.bz2
-cohere-documents-08-1k.json
-cohere-documents-08-1k.json.bz2
-cohere-documents-08.json
-cohere-documents-08.json.bz2
-cohere-documents-09-1k.json
-cohere-documents-09-1k.json.bz2
-cohere-documents-09.json
-cohere-documents-09.json.bz2
-cohere-documents-10-1k.json
-cohere-documents-10-1k.json.bz2
-cohere-documents-10.json
-cohere-documents-10.json.bz2
-cohere-documents-11-1k.json
-cohere-documents-11-1k.json.bz2
-cohere-documents-11.json
-cohere-documents-11.json.bz2
-cohere-documents-12-1k.json
-cohere-documents-12-1k.json.bz2
-cohere-documents-12.json
-cohere-documents-12.json.bz2
-cohere-documents-13-1k.json
-cohere-documents-13-1k.json.bz2
-cohere-documents-13.json
-cohere-documents-13.json.bz2
-cohere-documents-14-1k.json
-cohere-documents-14-1k.json.bz2
-cohere-documents-14.json
-cohere-documents-14.json.bz2
-cohere-documents-15-1k.json
-cohere-documents-15-1k.json.bz2
-cohere-documents-15.json
-cohere-documents-15.json.bz2
-cohere-documents-16-1k.json
-cohere-documents-16-1k.json.bz2
-cohere-documents-16.json
-cohere-documents-16.json.bz2
-cohere-documents-17-1k.json
-cohere-documents-17-1k.json.bz2
-cohere-documents-17.json
-cohere-documents-17.json.bz2
-cohere-documents-18-1k.json
-cohere-documents-18-1k.json.bz2
-cohere-documents-18.json
-cohere-documents-18.json.bz2
-cohere-documents-19-1k.json
-cohere-documents-19-1k.json.bz2
-cohere-documents-19.json
-cohere-documents-19.json.bz2
-cohere-documents-20-1k.json
-cohere-documents-20-1k.json.bz2
-cohere-documents-20.json
-cohere-documents-20.json.bz2
-cohere-documents-21-1k.json
-cohere-documents-21-1k.json.bz2
-cohere-documents-21.json
-cohere-documents-21.json.bz2
-cohere-documents-22-1k.json
-cohere-documents-22-1k.json.bz2
-cohere-documents-22.json
-cohere-documents-22.json.bz2
-cohere-documents-23-1k.json
-cohere-documents-23-1k.json.bz2
-cohere-documents-23.json
-cohere-documents-23.json.bz2
-cohere-documents-24-1k.json
-cohere-documents-24-1k.json.bz2
-cohere-documents-24.json
-cohere-documents-24.json.bz2
-cohere-documents-25-1k.json
-cohere-documents-25-1k.json.bz2
-cohere-documents-25.json
-cohere-documents-25.json.bz2
-cohere-documents-26-1k.json
-cohere-documents-26-1k.json.bz2
-cohere-documents-26.json
-cohere-documents-26.json.bz2
-cohere-documents-27-1k.json
-cohere-documents-27-1k.json.bz2
-cohere-documents-27.json
-cohere-documents-27.json.bz2
-cohere-documents-28-1k.json
-cohere-documents-28-1k.json.bz2
-cohere-documents-28.json
-cohere-documents-28.json.bz2
-cohere-documents-29-1k.json
-cohere-documents-29-1k.json.bz2
-cohere-documents-29.json
-cohere-documents-29.json.bz2
-cohere-documents-30-1k.json
-cohere-documents-30-1k.json.bz2
-cohere-documents-30.json
-cohere-documents-30.json.bz2
-cohere-documents-31-1k.json
-cohere-documents-31-1k.json.bz2
-cohere-documents-31.json
-cohere-documents-31.json.bz2
-cohere-documents-32-1k.json
-cohere-documents-32-1k.json.bz2
-cohere-documents-32.json
-cohere-documents-32.json.bz2
-cohere-documents-33-1k.json
-cohere-documents-33-1k.json.bz2
-cohere-documents-33.json
-cohere-documents-33.json.bz2
-cohere-documents-34-1k.json
-cohere-documents-34-1k.json.bz2
-cohere-documents-34.json
-cohere-documents-34.json.bz2
-cohere-documents-35-1k.json
-cohere-documents-35-1k.json.bz2
-cohere-documents-35.json
-cohere-documents-35.json.bz2
-cohere-documents-36-1k.json
-cohere-documents-36-1k.json.bz2
-cohere-documents-36.json
-cohere-documents-36.json.bz2
-cohere-documents-37-1k.json
-cohere-documents-37-1k.json.bz2
-cohere-documents-37.json
-cohere-documents-37.json.bz2
-cohere-documents-38-1k.json
-cohere-documents-38-1k.json.bz2
-cohere-documents-38.json
-cohere-documents-38.json.bz2
-cohere-documents-39-1k.json
-cohere-documents-39-1k.json.bz2
-cohere-documents-39.json
-cohere-documents-39.json.bz2
-cohere-documents-40-1k.json
-cohere-documents-40-1k.json.bz2
-cohere-documents-40.json
-cohere-documents-40.json.bz2
-cohere-documents-41-1k.json
-cohere-documents-41-1k.json.bz2
-cohere-documents-41.json
-cohere-documents-41.json.bz2
-cohere-documents-42-1k.json
-cohere-documents-42-1k.json.bz2
-cohere-documents-42.json
-cohere-documents-42.json.bz2
-cohere-documents-43-1k.json
-cohere-documents-43-1k.json.bz2
-cohere-documents-43.json
-cohere-documents-43.json.bz2
-cohere-documents-44-1k.json
-cohere-documents-44-1k.json.bz2
-cohere-documents-44.json
-cohere-documents-44.json.bz2
-cohere-documents-45-1k.json
-cohere-documents-45-1k.json.bz2
-cohere-documents-45.json
-cohere-documents-45.json.bz2
-cohere-documents-46-1k.json
-cohere-documents-46-1k.json.bz2
-cohere-documents-46.json
-cohere-documents-46.json.bz2
-cohere-documents-47-1k.json
-cohere-documents-47-1k.json.bz2
-cohere-documents-47.json
-cohere-documents-47.json.bz2
