--- cohere_v3_vector/operations/default.json
@@ -1,41 +1,0 @@
-{
-  "name": "create-index",
-  "operation-type": "create-index",
-  "settings": {{index_settings | default({}) | tojson}}
-},
-{
-  "name": "check-cluster-health",
-  "operation-type": "cluster-health",
-  "request-params": {
-    "wait_for_status": "green"
-  },
-  "retry-until-success": true
-},
-{
-  "name": "initial-documents-indexing",
-  "operation-type": "bulk",
-  "corpora": "cohere-v3-initial-indexing",
-  "bulk-size": {{initial_indexing_bulk_size | default(500)}},
-  "ingest-percentage": {{initial_indexing_ingest_percentage | default(100)}}
-},
-{
-  "name": "parallel-documents-indexing",
-  "operation-type": "bulk",
-  "corpora": "cohere-v3-parallel-indexing",
-  "bulk-size": {{parallel_indexing_bulk_size | default(500)}},
-  "ingest-percentage": {{parallel_indexing_ingest_percentage | default(100)}}
-},
-{
-  "name": "knn-search-10-100",
-  "operation-type": "search",
-  "param-source": "knn-param-source",
-  "k": 10,
-  "num-candidates": 100
-},
-{
-  "name": "knn-search-100-1000",
-  "operation-type": "search",
-  "param-source": "knn-param-source",
-  "k": 100,
-  "num-candidates": 1000
-}
