+++ msmarco-v2-vector/.gitignore
@@ -1,0 +1,1 @@
+queries.json
