+++ msmarco-v2-vector/_tools/requirements.txt
@@ -1,0 +1,3 @@
+cohere
+datasets
+ir-datasets
